+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.2.0]
+* add badge support to the sidemenu
+* change IconData to Icon for more flexibility
+
 ## [0.1.1+1]
 * add demo to readme